requirements.txt
alembic==1.8.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
numpy==1.23.1
opencv-python==4.6.0.66
picamera==1.13
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
SQLAlchemy==1.4.39
twilio==7.11.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
